requirements.txt
django
djangorestframework
django-import-export
psycopg2-binary
python-dotenv
django-compressor
pandas
openpyxl